refactor(cli): split identity module into file
